build: Add setup and initial build files
These files need to be copied (or linked) to the QEMU folder.

The setup script runs all the required steps to correctly export the
files and configure the QEMU build process.
It needs to be run before building QEMU with the new device.

Source file is only stub for the moment.